Fixes #96, #101, #99 and belbio/bel_sources#96
Fixed issue with equivalencing - both the issue of slow equivalencing due to using up to 10 steps for equivalences and the issue that alt_ids were not included in equivalencing.

Also fixed the issue with dashes and underscores being used to tokenize text for term completion - now we only use colons and spaces for tokenization.

Also shifted api routes into a separate file.